+++ src/icemac/ab/importer/conftest.py
@@ -5,8 +5,8 @@ import icemac.addressbook.testing
 import pytest
 
 
-pytest_plugins = 'icemac.addressbook.conftest'
 
+pytest_plugins = 'icemac.addressbook.fixtures'
 
 # Fixtures to set-up infrastructure which are usable in tests:
 
@@ -14,7 +14,7 @@ pytest_plugins = 'icemac.addressbook.conftest'
 @pytest.yield_fixture(scope='function')
 def address_book(addressBookConnectionF):
     """Get the address book with importer as site."""
-    for address_book in icemac.addressbook.conftest.site(
+    for address_book in icemac.addressbook.testing.site(
             addressBookConnectionF):
         yield address_book
 
@@ -22,7 +22,7 @@ def address_book(addressBookConnectionF):
 @pytest.fixture(scope='function')
 def browser(browserWsgiAppS):
     """Fixture for testing with zope.testbrowser."""
-    assert icemac.addressbook.conftest.CURRENT_CONNECTION is not None, \
+    assert icemac.addressbook.testing.CURRENT_CONNECTION is not None, \
         "The `browser` fixture needs a database fixture like `address_book`."
     return icemac.ab.importer.testing.Browser(wsgi_app=browserWsgiAppS)
 
@@ -41,23 +41,44 @@ def ImportFileFactory(FileFactory):
     return create_file
 
 
+# Fixtures to help asserting
+
+@pytest.fixture(scope='function')
+def sitemenu(browser):
+    """Helper fixture to test the selections in the site menu."""
+    return icemac.addressbook.testing.SiteMenu
+
+
+@pytest.fixture(scope='function')
+def assert_address_book(address_book):
+    """Fixture returning an object providing a custom address book asserts."""
+    return icemac.addressbook.testing.AddressBookAssertions(address_book)
+
+
 # Infrastructure fixtures
 
 
 @pytest.yield_fixture(scope='session')
-def zcmlS(zcmlS):
+def zcmlS():
     """Load importer ZCML on session scope."""
     layer = icemac.addressbook.testing.SecondaryZCMLLayer(
-        'Importer', __name__, icemac.ab.importer, [zcmlS])
+        'Importer', __name__, icemac.ab.importer)
     layer.setUp()
     yield layer
     layer.tearDown()
 
 
 @pytest.yield_fixture(scope='session')
+def zodbS(zcmlS):
+    """Create an empty test ZODB."""
+    for zodb in icemac.addressbook.testing.pyTestEmptyZodbFixture():
+        yield zodb
+
+
+@pytest.yield_fixture(scope='session')
 def addressBookS(zcmlS, zodbS):
     """Create an address book for the session."""
-    for zodb in icemac.addressbook.conftest.pyTestAddressBookFixture(
+    for zodb in icemac.addressbook.testing.pyTestAddressBookFixture(
             zodbS, 'ImporterS'):
         yield zodb
 
@@ -65,6 +86,6 @@ def addressBookS(zcmlS, zodbS):
 @pytest.yield_fixture(scope='function')
 def addressBookConnectionF(addressBookS):
     """Get the connection to the right demo storage."""
-    for connection in icemac.addressbook.conftest.pyTestStackDemoStorage(
+    for connection in icemac.addressbook.testing.pyTestStackDemoStorage(
             addressBookS, 'ImporterF'):
         yield connection